Create incompleteGUI.py
This is my test to see how we can use tkinter to write the GUI. 
A bug that took me two hours but still hasn't been solved is the radiobuttons do not have a default selection.
When hovering over "Always, Ask, Never", all three appear to be selected. Once one is selected, the problem does not show up again.